requirements.txt
opencv-python
cvzone
streamlit
google-generativeai
Pillow
numpy
mediapipe

